+++ test/test_helper.rb
@@ -1,0 +1,3 @@
+require 'minitest/autorun'
+
+Dir.glob('./test/**/*_spec.rb', &method(:require)) if __FILE__ == $0