+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v0.4.0
+
+* leaving empty values out when converting to a hash was too intrusive.
+  If you need that, do it in your own implementation, it's trivial
+
 ### v0.3.0
 
 * it can be instantiated with no arguments